mv jobs_where_path_exists to the prepare module
