+++ photon_stream/production/tools.py
@@ -1,14 +1,3 @@
-import os
-import numpy as np
-import json
-import datetime as dt
-
-def jobs_where_path_exists(jobs, path='raw_path'):
-    accesible_jobs = []
-    for job in jobs:
-        if os.path.exists(job[path]):
-            accesible_jobs.append(job)
-    return accesible_jobs
 
 
 def night_id_2_yyyy(night):
